refactoring gig to gig_handlers
